Make sure that shel provisioning is not made more than once  - added check for /var/vagrant_provision to provision.sh

+++ provisioning/provision.sh
@@ -1,5 +1,12 @@
 echo "Installing Apache and setting up link to /vagrant"
+
+if [ -f "/var/vagrant_provision" ] then
+  exit(0)
+if
+
 apt-get update >/dev/null 2>&1
 apt-get install -y apache2 >/dev/null 2>&1
 rm -rf /var/www
-ln -fs /vagranto/work /var/www
+ln -fs /vagrant/work /var/www
+
+touch /var/vagrant_provision
